refactor: reveal question api and removed answer page flow from claim questions
