+++ app/api/question/reveal/route.ts
@@ -1,17 +1,10 @@
-import {
-  getDeckThatNeedChompResultBasedOnRevealedQuestionIds,
-  handleFirstRevealToPopulateSubjectiveQuestion,
-} from "@/app/actions/chompResult";
-import { handleSendBonk } from "@/app/actions/claim";
 import { MINIMAL_ANSWER_COUNT } from "@/app/constants/answers";
+import {
+  claimAllSelected,
+  revealAllSelected,
+} from "@/app/queries/bot/processBurnClaim";
 import { queryQuestionsForReadyToReveal } from "@/app/queries/home";
-import { questionAnswerCountQuery } from "@/app/queries/questionAnswerCountQuery";
 import prisma from "@/app/services/prisma";
-import { calculateReward } from "@/app/utils/algo";
-import { ONE_MINUTE_IN_MILLISECONDS } from "@/app/utils/dateUtils";
-import { calculateRevealPoints } from "@/app/utils/points";
-import { isEntityRevealable } from "@/app/utils/question";
-import { FungibleAsset, ResultType, TransactionStatus } from "@prisma/client";
 import { headers } from "next/headers";
 
 export async function GET(req: Request) {
@@ -36,8 +29,6 @@ export async function GET(req: Request) {
       question.answerCount && question.answerCount >= MINIMAL_ANSWER_COUNT,
   );
 
-  // const getQuestionIds = () => questionMinAnswerFilter.map((item) => item.id);
-
   return Response.json(questionMinAnswerFilter);
 }
 
@@ -56,254 +47,28 @@ export async function POST(req: Request) {
     return Response.json("userId parameter is required", { status: 400 });
   }
 
+  const userWallet = await prisma.wallet.findFirst({
+    where: {
+      userId: userId,
+    },
+  });
+
+  if (!userWallet) {
+    return Response.json("No wallet found for the user", { status: 400 });
+  }
+
   const data = await req.json();
   const { questionIds, burnTx } = data;
 
-  // const release = await acquireMutex({
-  //   identifier: "REVEAL",
-  //   data: { userId: userId },
-  // });
-
   try {
-    const userWallet = await prisma.wallet.findFirst({
-      where: {
-        userId: userId,
-      },
-    });
-
-    if (!userWallet) {
-      return;
-    }
-    await handleFirstRevealToPopulateSubjectiveQuestion(questionIds);
-    const questionsFilteredForUser = await prisma.question.findMany({
-      where: {
-        id: { in: questionIds },
-      },
-      include: {
-        chompResults: {
-          where: {
-            userId: userId,
-          },
-        },
-      },
-    });
-    const questionAnswersCount = await questionAnswerCountQuery(questionIds);
-
-    const revealableQuestions = questionsFilteredForUser.filter((question) =>
-      isEntityRevealable({
-        revealAtAnswerCount: question.revealAtAnswerCount,
-        revealAtDate: question.revealAtDate,
-        answerCount:
-          questionAnswersCount.find((qa) => qa.id === question.id)?.count ?? 0,
-      }),
-    );
-
-    if (!revealableQuestions.length) {
-      throw new Error("No revealable questions available");
-    }
-
-    const revealableQuestionIds = revealableQuestions.map((q) => q.id);
-    const decksToAddRevealFor =
-      await getDeckThatNeedChompResultBasedOnRevealedQuestionIds(
-        revealableQuestionIds,
-        userId,
-      );
-
-    const rewardsPerQuestionId = await calculateReward(
+    await revealAllSelected(questionIds, userId, burnTx);
+    const results = await claimAllSelected(
+      questionIds,
       userId,
-      revealableQuestionIds,
+      userWallet?.address,
     );
-
-    const revealPoints = await calculateRevealPoints(
-      Object.values(rewardsPerQuestionId!),
-    );
-
-    const pointsAmount = revealPoints.reduce((acc, cur) => acc + cur.amount, 0);
-
-    await prisma.$transaction(async (tx) => {
-      await tx.chompResult.deleteMany({
-        where: {
-          AND: {
-            userId: userId,
-            questionId: {
-              in: revealableQuestionIds,
-            },
-            burnTransactionSignature: burnTx,
-            transactionStatus: TransactionStatus.Pending,
-          },
-        },
-      });
-
-      await tx.chompResult.createMany({
-        data: [
-          ...revealableQuestionIds.map((questionId) => ({
-            questionId,
-            userId: userId,
-            result: ResultType.Revealed,
-            burnTransactionSignature: burnTx,
-            rewardTokenAmount: rewardsPerQuestionId?.[questionId],
-            transactionStatus: TransactionStatus.Completed,
-          })),
-          ...decksToAddRevealFor.map((deck) => ({
-            deckId: deck.id,
-            userId: userId,
-            result: ResultType.Revealed,
-            burnTransactionSignature: burnTx,
-          })),
-        ],
-      });
-
-      await tx.fungibleAssetBalance.upsert({
-        where: {
-          asset_userId: {
-            asset: FungibleAsset.Point,
-            userId: userId,
-          },
-        },
-        update: {
-          amount: {
-            increment: pointsAmount,
-          },
-        },
-        create: {
-          userId: userId,
-          asset: FungibleAsset.Point,
-          amount: pointsAmount,
-        },
-      });
-
-      const campaignId = revealableQuestions[0].campaignId;
-
-      if (!!campaignId) {
-        const currentDate = new Date();
-
-        await tx.dailyLeaderboard.upsert({
-          where: {
-            user_campaign_date: {
-              userId: userId,
-              campaignId: revealableQuestions[0].campaignId!,
-              date: currentDate,
-            },
-          },
-          create: {
-            userId: userId,
-            campaignId: revealableQuestions[0].campaignId,
-            points: pointsAmount,
-          },
-          update: {
-            points: { increment: pointsAmount },
-          },
-        });
-      }
-
-      await tx.fungibleAssetTransactionLog.createMany({
-        data: revealPoints.map((revealPointsTx) => ({
-          asset: FungibleAsset.Point,
-          type: revealPointsTx.type,
-          change: revealPointsTx.amount,
-          userId: userId,
-        })),
-      });
-    });
-
-    const chompResults = await prisma.chompResult.findMany({
-      where: {
-        userId: userId,
-        questionId: {
-          in: questionIds,
-        },
-        result: ResultType.Revealed,
-      },
-    });
-    const sendTx = await handleSendBonk(chompResults, userWallet.address);
-
-    await prisma.$transaction(
-      async (tx) => {
-        await tx.chompResult.updateMany({
-          where: {
-            id: {
-              in: chompResults.map((r) => r.id),
-            },
-          },
-          data: {
-            sendTransactionSignature: sendTx,
-          },
-        });
-
-        const questionIdsClaimed = chompResults.map((cr) => cr.questionId ?? 0);
-        // We're querying dirty data in transaction so we need claimed questions
-        const decks = await tx.deck.findMany({
-          where: {
-            deckQuestions: {
-              some: {
-                questionId: {
-                  in: questionIdsClaimed,
-                },
-              },
-              every: {
-                question: {
-                  chompResults: {
-                    every: {
-                      userId: userId,
-                      result: ResultType.Claimed,
-                    },
-                  },
-                },
-              },
-            },
-          },
-          include: {
-            chompResults: {
-              where: {
-                userId: userId,
-              },
-            },
-          },
-        });
-
-        const deckRevealsToUpdate = decks
-          .filter((deck) => deck.chompResults && deck.chompResults.length > 0)
-          .map((deck) => deck.id);
-
-        if (deckRevealsToUpdate.length > 0) {
-          await tx.chompResult.updateMany({
-            where: {
-              deckId: { in: deckRevealsToUpdate },
-            },
-            data: {
-              result: ResultType.Claimed,
-              sendTransactionSignature: sendTx,
-            },
-          });
-        }
-
-        const deckRevealsToCreate = decks
-          .filter(
-            (deck) => !deck.chompResults || deck.chompResults.length === 0,
-          )
-          .map((deck) => deck.id);
-
-        if (deckRevealsToCreate.length > 0) {
-          await tx.chompResult.createMany({
-            data: deckRevealsToCreate.map((deckId) => ({
-              deckId,
-              userId: userId,
-              result: ResultType.Claimed,
-              sendTransactionSignature: sendTx,
-            })),
-          });
-        }
-      },
-      {
-        isolationLevel: "Serializable",
-        timeout: ONE_MINUTE_IN_MILLISECONDS,
-      },
-    );
-    return Response.json(chompResults);
-    // release();
-    // revalidatePath("/application");
+    return Response.json(results?.chompResults);
   } catch (e) {
-    // release();
     console.log(e);
     return Response.json("Error at reveal", { status: 400 });
   }
